+++ src/Domain.LinnApps/GoodsIn/IGoodsInService.cs
@@ -11,7 +11,7 @@
             int createdBy,
             string partNumber,
             string manufacturersPartNumber,
-            int qty,
+            decimal qty,
             int? orderNumber,
             int? orderLine,
             int? loanNumber,
